Separated: Admin setting field logic into it's own field classes. Eventualy need to make a fields class we can extend from.
